XRDDEV-584 & 681 Access rights PR comment fixes part 7
* code cleanup
