+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/service/AccessRightService.java
@@ -44,7 +44,6 @@ import org.niis.xroad.restapi.dto.AccessRightHolderDto;
 import org.niis.xroad.restapi.exceptions.ErrorDeviation;
 import org.niis.xroad.restapi.facade.GlobalConfFacade;
 import org.niis.xroad.restapi.repository.ClientRepository;
-import org.niis.xroad.restapi.repository.IdentifierRepository;
 import org.niis.xroad.restapi.repository.LocalGroupRepository;
 import org.niis.xroad.restapi.util.FormatUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -53,7 +52,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -76,18 +74,18 @@ public class AccessRightService {
     private final GlobalConfFacade globalConfFacade;
     private final ClientRepository clientRepository;
     private final ServiceService serviceService;
-    private final IdentifierRepository identifierRepository;
+    private final IdentifierService identifierService;
     private final GlobalConfService globalConfService;
 
     @Autowired
     public AccessRightService(LocalGroupRepository localGroupRepository, GlobalConfFacade globalConfFacade,
-            ClientRepository clientRepository, ServiceService serviceService, IdentifierRepository identifierRepository,
+            ClientRepository clientRepository, ServiceService serviceService, IdentifierService identifierService,
             GlobalConfService globalConfService) {
         this.localGroupRepository = localGroupRepository;
         this.globalConfFacade = globalConfFacade;
         this.clientRepository = clientRepository;
         this.serviceService = serviceService;
-        this.identifierRepository = identifierRepository;
+        this.identifierService = identifierService;
         this.globalConfService = globalConfService;
     }
 
@@ -288,36 +286,6 @@ public class AccessRightService {
     }
 
     /**
-     * Verify that all identifiers are authentic, then get the existing ones from the local db and persist
-     * the not-existing ones. This is a necessary step if we are changing identifier relations (such as adding
-     * access rights to services)
-     * @param xRoadIds {@link GlobalGroupId} or {@link ClientId}
-     * @return List of XRoadIds ({@link GlobalGroupId} or {@link ClientId})
-     * @throws IdentifierNotFoundException
-     */
-    private Set<XRoadId> getOrPersistSubsystemAndGlobalGroupIds(Set<XRoadId> xRoadIds)
-            throws IdentifierNotFoundException {
-        // Check that the identifiers exist in globalconf
-        // LocalGroups must be verified separately! (they do not exist in globalconf)
-        Set<XRoadId> subsystemsAndGlobalGroups = xRoadIds.stream()
-                .filter(xRoadId -> xRoadId.getObjectType() == XRoadObjectType.SUBSYSTEM
-                        || xRoadId.getObjectType() == XRoadObjectType.GLOBALGROUP)
-                .collect(Collectors.toSet());
-        if (!globalConfService.identifiersExist(subsystemsAndGlobalGroups)) {
-            // This exception should be pretty rare since it only occurs if bogus subjects are found
-            throw new IdentifierNotFoundException();
-        }
-        Collection<XRoadId> allIdsFromDb = identifierRepository.getIdentifiers();
-        Set<XRoadId> txEntities = allIdsFromDb.stream()
-                .filter(xRoadIds::contains) // this works because of the XRoadId equals and hashCode overrides
-                .collect(Collectors.toSet());
-        xRoadIds.removeAll(txEntities); // remove the persistent ones
-        identifierRepository.saveOrUpdate(xRoadIds); // persist the non-persisted
-        txEntities.addAll(xRoadIds); // add the newly persisted ids into the collection of already existing ids
-        return txEntities;
-    }
-
-    /**
      * Adds access rights to services
      * @param clientId
      * @param fullServiceCode
@@ -338,6 +306,28 @@ public class AccessRightService {
         Set<XRoadId> localGroups = getLocalGroupsAsXroadIds(localGroupIds);
         txSubjects.addAll(localGroups);
         return addServiceAccessRights(clientId, fullServiceCode, txSubjects);
+    }
+
+    /**
+     * Verify that all identifiers are authentic, then get the existing ones from the local db and persist
+     * the not-existing ones. This is a necessary step if we are changing identifier relations (such as adding
+     * access rights to services)
+     * @param subjectIds {@link GlobalGroupId} or {@link ClientId}
+     * @return List of XRoadIds ({@link GlobalGroupId} or {@link ClientId})
+     */
+    private Set<XRoadId> getOrPersistSubsystemAndGlobalGroupIds(Set<XRoadId> subjectIds)
+            throws IdentifierNotFoundException {
+        // Check that the identifiers exist in globalconf
+        // LocalGroups must be verified separately! (they do not exist in globalconf)
+        Set<XRoadId> subsystemsAndGlobalGroups = subjectIds.stream()
+                .filter(xRoadId -> xRoadId.getObjectType() == XRoadObjectType.SUBSYSTEM
+                        || xRoadId.getObjectType() == XRoadObjectType.GLOBALGROUP)
+                .collect(Collectors.toSet());
+        if (!globalConfService.identifiersExist(subsystemsAndGlobalGroups)) {
+            // This exception should be pretty rare since it only occurs if bogus subjects are found
+            throw new IdentifierNotFoundException();
+        }
+        return identifierService.getOrPersistXroadIds(subsystemsAndGlobalGroups);
     }
 
     /**
@@ -368,17 +358,6 @@ public class AccessRightService {
             super(new ErrorDeviation(ERROR_ACCESSRIGHT_NOT_FOUND));
         }
 
-    }
-
-    /**
-     * If identifier was not found
-     */
-    public static class IdentifierNotFoundException extends NotFoundException {
-        public static final String ERROR_IDENTIFIER_NOT_FOUND = "identifier_not_found";
-
-        public IdentifierNotFoundException() {
-            super(new ErrorDeviation(ERROR_IDENTIFIER_NOT_FOUND));
-        }
     }
 
     /**
